chore: split up fixture handlers
