+++ packages/skylarktv/src/mocks/fixtures/listGenres.ts
@@ -1,0 +1,23 @@
+import { graphql, HttpResponse } from "msw";
+import { SAAS_API_ENDPOINT } from "../../constants/env";
+import { airtableData } from "../airtableData";
+
+export const listGenresHandlers = [
+  // Handle Genre list
+  graphql.link(SAAS_API_ENDPOINT).query("LIST_GENRES", ({ variables }) => {
+    const genres = (airtableData.genres || []).map((genre) => ({
+      uid: genre.id,
+      name: genre.fields.name,
+      slug: genre.fields.slug,
+    }));
+
+    return HttpResponse.json({
+      data: {
+        listObjects: {
+          next_token: null,
+          objects: genres,
+        },
+      },
+    });
+  }),
+];
