+++ packages/skylarktv/src/mocks/fixtures/skylarktvConfig.ts
@@ -1,0 +1,61 @@
+import { graphql, HttpResponse } from "msw";
+import { SAAS_API_ENDPOINT } from "../../constants/env";
+import { getObjectsByType } from "../airtableData";
+
+// Mock SkylarkTV Config
+const mockSkylarkTVConfig = {
+  getSkylarkSet: {
+    content: {
+      objects: [
+        {
+          uid: "config-1",
+          __typename: "SkylarkLiveAsset",
+          title: "Skylark TV Config",
+          hls_url: null,
+          url: "https://skylark-references-production.s3.eu-west-1.amazonaws.com/skylarktv/configs/default-config.json",
+        },
+      ],
+    },
+  },
+};
+
+// Create homepage set from Airtable data
+const createHomepageSet = () => {
+  const movies = getObjectsByType("movies").slice(0, 10);
+  const episodes = getObjectsByType("episodes").slice(0, 10);
+  const brands = getObjectsByType("brands").slice(0, 5);
+
+  return {
+    uid: "homepage-set",
+    title: "Homepage",
+    content: {
+      objects: [...movies, ...episodes, ...brands],
+    },
+  };
+};
+
+export const skylarktvConfigHandlers = [
+  graphql.link(SAAS_API_ENDPOINT).query("GET_SKYLARK_TV_CONFIG", () =>
+    HttpResponse.json({
+      data: mockSkylarkTVConfig,
+    })
+  ),
+
+  graphql.link(SAAS_API_ENDPOINT).query("GET_APP_CONFIG", () => {
+    const homepageSet = createHomepageSet();
+    
+    return HttpResponse.json({
+      data: {
+        getSkylarkSet: homepageSet,
+      },
+    });
+  }),
+
+  graphql.link(SAAS_API_ENDPOINT).query("GET_STREAMTV_CONFIG", () => {
+    return HttpResponse.json({
+      data: {
+        getSkylarkSet: createHomepageSet(),
+      },
+    });
+  }),
+];
